Alteração de tipo decimal para inteiro.

+++ ContribuaMais.API/Models/Dtos/DtoCampanha.cs
@@ -6,7 +6,7 @@
         public DateTime Inicio { get; set; }
         public DateTime Fim { get; set; }
         public bool EstaAtiva { get; set; }
-        public decimal Objetivo { get; set; }
-        public decimal Arrecadado { get; set; }
+        public int Objetivo { get; set; }
+        public int Arrecadado { get; set; }
     }
 }
